Add proper error when build dir and tap dir are not sync
As explained in https://docs.travis-ci.com/user/environment-variables,

TRAVIS_COMMIT_RANGE: The range of commits that were included in the push or
pull request. (Note that this is empty for builds triggered by the initial
commit of a new branch.)

In the case where TRAVIS_COMMIT_RANGE is an empty string, we want
to avoid the 'nil to String' conversion error triggered because
diff_start_sha1 and diff_end_sha1 are nil.

Here is the error that would be displayed:

ARGV: --verbose --no-pull --coverage
Error: no implicit conversion of nil into String
-> cmd/brew-test-bot.rb:452:in `download'