when pushing the bottle commit to the <git-name>'s fork, only push tag (BREAKING)
On the Homebrew-core tap, I suppose that the purpose of pushing the master ref is
to reflect the latest pushes on the master branch, so that visitors of BrewTestBot
can see that BrewTestBot is in use. The problem with this is that I use the same
repo for my tap and for pushing tags (e.g., pr-123), which makes the force push
to master a disaster.

Maybe we could add a new flag that would enable the force push of master. For now,
though, I disable it altogether.